requirements.txt
gradio>=3.0
youtube-transcript-api>=0.5.0
openai>=0.27.0
google-api-python-client>=2.70.0
google-auth>=2.17.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
